Rework how packages are imported - see #590 & #587
For the most part, it is still done the same.

The strategy has changed because the reporting mechanism for new packages can
work on only one ABI/set at a time. This is not compatible with how the script
was being used. The script could be updated, but I think that would
unnecessarily complicate it.

Instead, I have slightly altered the approach.

It was:

* check for new available repos
* import them into the packages_new table
* update the packages table with the new information
* send out notification emails based on the package_notifications table

It will be:

* check for new available repos
* for each repo
  * import it into packages_new table
  * update the packages table with the new information
  * send out notification emails based on the package_notifications table

The rest of this commit message will describe changes to each file.

README.md

* Update based on the above changes

UpdatePackagesFromRawPackages.py

* stop clearing the SIGNAL_NEW_REPO_IMPORTED signal - this signal is no longer
  required

fetch-extract-parse-import-one-abi.sh

* Add more logging

import-via-copy-packagesite-all-raw-fields.py

* remove duplicate syslog.openlog & syslog.syslog entries from top of script
* after importing a repo run UpdatePackagesFromRawPackages.py and then
  report-notification-packages.pl
* stop setting SIGNAL_NEW_REPO_IMPORTED because we already do the work
  associated with that signal